requirements.txt
streamlit
tensorflow
opencv-python-headless
pillow
numpy
pytesseract
mysql-connector-python
